requirements.txt
torch
torchvision
timm
transformers
fastapi
uvicorn
streamlit
gradio
pillow
numpy
captum
datasets